Update Sidebar design
