+++ src/containers/tutorial/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import './styles.css';
+
+const Tutorial = ()=> {
+    return (
+        <div>
+            wer
+        </div>
+    );
+};
+
+export default Tutorial;
