+++ src/views/playground/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import Sidebar from "../../containers/sidebar";
 import Content from "../../containers/content";
 import {RootProvider} from "../../context/RootProvider";
 import './styles.css';
+import Sidebar from "../../containers/sidebar";
 
 const Playground = ()=> {
     return (
